Dont eliminate resolution on shapes
Summary:
It appears that we cannot eliminate resolution on shapes since resolution phase removes "value" field in a shape's type structure. On this diff, I killed the code to remove the resolution.
A better future solution would be implement this removal of "value" field in `resolveTSStatically` function but currently it doesn't do anything for shapes.

Reviewed By: jjgriego

Differential Revision: D19280425

fbshipit-source-id: 68bd5db8fb089626faa98f13a5c08f318608c131