Type accesses cleanup: remove choose_assigned_type and add comments
Summary:
This diff greatly improves the readability of Typing_taccess. It finishes
to remove the local env type and replaces it with a well-documented context
record which only flows down the call stack. The two roles of
choose_assigned_type are now replaced by more understandable mechanisms
(Exact propagation and a concept of base of the expansion).

See the tasks linked for the type holes I discovered while reworking this code.

Reviewed By: andrewjkennedy

Differential Revision: D19162308

fbshipit-source-id: ab12cfd0b078fac13039ea217caa9c017bb11e6d