Correctly skip generating a stub for the dependency on the target method
Summary: We incorrectly test if a given `Method` or `SMethod` dependency is on the target method that is being extracted: we only check if the method names are the same, and we don't check if the class names are also the same, which results in us not generating stubs for methods whose names match the target method name.  One example where this manifests itself is when the target method has an `<<__Override>>` attribute: in this case, we don't generate the parent, overridden method, which results in a type error in the extracted code.

Reviewed By: Wilfred

Differential Revision: D19300707

fbshipit-source-id: 13499159a31923d157f13d0fb7465892a1308e1b