Crude hack to support XHP
Summary:
Because of namespaces, the extraction tool extracts class names fully qualified with the namespace path.  Unfortunately, for XHP components this leads to names like `\:foo`, which are not even accepted by Hack parser.  (Fun fact: one can define an XHP component in a namespace but cannot use it outside of that namespace because there is no way to refer to it.)

Ideally, we should fix XHP and the way it interacts with namespaces.  To support XHP in the extraction tool, I'm proposing a really crude hack, which is to do a textual search of substrings `\:` in the output and replace them with `:` before formatting.

Reviewed By: Wilfred

Differential Revision: D19330992

fbshipit-source-id: 03645740ceb1a5f2248b640e17504baef400a470