deprecate is_hh_file
Summary:
`is_hh_file` is true if current file is not `file_info::Mphp` and false if it is `file_info::Mphp`.

This boolean value is passed to Ocaml lowerer and returned it as lowerer result. The value can be inferred from file mode.

`is_hh_file` is only used in compile.ml.

`hh_parse` provides a way to overwrite it, it doesn't seem it is used any where. If overwriting is needed, it should directly set the value where it is consumed, don't pass it to lowerer.

bypass-lint

Differential Revision: D19234280

fbshipit-source-id: ec5e39d9f0362b1352a72585e1ba0529caeb181e