add folly noexcept patch (#8641)
Summary:
As discussed on D19351227, this implementation doesn't correctly handle class member function pointers, but it looks like nothing in the HHVM build depends on that so we can use it to fix the nightly builds.

I'll remove this once we have a proper implementation in Folly.
Pull Request resolved: https://github.com/facebook/hhvm/pull/8641

Differential Revision: D19353214

Pulled By: jjergus

fbshipit-source-id: b47a439d3e21424050162602c07311acf1e3c15c