preserve varray/darrays in the APCPrime upgrade codepath
Summary:
The apc_prime.so library we load at runtime is not the exact one we first produce in the build. There is an intermediate upgrade step, which looks something like:
```
hhvm -m server -vServer.APC.EnableConstLoad=true -vServer.APC.PrimeLibrary="$APC_PRIME_OUTPUT" \
  -vServer.APC.PrimeLibraryUpgradeDest="$APC_PRIME_OUTPUT".opt -vServer.APC.FileStorage.Enable=true \
  -vServer.AllowRunAsRoot=true
```

This decompresses and interprets the raw .so we produced from cachearchiver output and serializes it to yet another custom format. That step was missing the `VARRAY_DARRAY` flag. I confirmed that the original inputs do properly have varray/darray serialized and that apc serialization at runtime properly preserves varray/darray-ness so this is (hopefully) the last missing link.

Reviewed By: viratyosin

Differential Revision: D19353397

fbshipit-source-id: cad3a68efa31e514b435555c8f1b20c4497cc4d1