clean up codemoder for parameters and returns, clean up options, enable tast tests
Summary:
There are options to infer only return types or only parameters which perform just a local inference. These were sort of warm ups for the global inference internship and are not very useful in themselves, so this diff cleans them up.

This diff also cleans up options for hh_single_type_check: there is now only `--global-inference`, the other options (`--infer-missing`, `--dump-tvenvs` and so on) are killed. To dump inference envs, use `--verbose 1` or `2`.

Reviewed By: manzyuk

Differential Revision: D19177809

fbshipit-source-id: 143f7440604e55401486c5ecd8dd6ef3f915e4a7