don't create global type vars for lambdas
Summary:
Global type variables were created for unannotated lambdas and anonymous function. That might have been all right, but most likely unnecessarily loading the global constraint graph.

This diff fixes that by wiring an `~is_lambda` flag all over the place. I'm not fond of this, another solution would be to turn the `f_name` field in `Aast.fun_` a `sid option` to be able to see when a `fun_` is anonymous, but that's a much bigger refactoring.

Differential Revision: D19371642

fbshipit-source-id: 8d7032d8681cae6fa2559eb8082dc9ca6c44bc4d