Sort errors correctly
Summary:
To avoid reporting duplicate errors, we sorted the errors by position (path + line #, start and end), then removed consecutive duplicates using the regular equality comparison.

This worked fairly well because even though we would still get duplicate errors this way, their number would be the same from run to run.

The reason we would still get duplicates is because at the same position, we could be reporting:
* errors with different error codes
* errors with the same error code but different warrants (e.g., class A fails to implement a method of interface I1 because it implements I2 and I3, both of which extend I1)

Facebook
This failure to correctly sort in all cases becomes an issue for the remote type checker because we may get the same errors both from the local and the remote workers due to work stealing.

The solution is simply to compare the rest of the properties if the positions are the same, and to compare all the messages and not just the first one.

Reviewed By: arxanas

Differential Revision: D19361356

fbshipit-source-id: 29f146543754669ae9549bac09865a54c9b47a82