Port error-sorting to Rust
Summary:
Write a method `Errors::into_sorted_vec` which matches the behavior of `Errors.get_sorted_error_list` in OCaml after {D19361356}.

A comment on Pos indicates that positions in different files should not be comparable, but I don't agree--it seems reasonable to compare positions in different files in use cases like this one (plus, we do it already in OCaml).

This diff also replaces the derived implementation of Ord and PartialOrd on FilePosSmall and FilePosLarge with an implementation matching the behavior of Pos::cmp (even though it is not used in Pos::cmp).

Reviewed By: jewelpit

Differential Revision: D19382274

fbshipit-source-id: 684c1c75ee8d2bcde24fe0b818732ddb1818f948