Fix couldBe operator in HHBBC around mismatched array data
Summary:
HHBBC incorrectly reports that types like PArr([Int]) and
PArr([Obj]) cannot could be each other. This is false, as their
intersection contains the empty array. The reason is because the
"couldBe" operator is always checking the data, and if they do not
overlap, returning false. We don't want to check the data if the bit
intersection contains the empty array because we already have a valid
intersection. We already deal with this properly for optional types,
so do the same for empty arrays. The only complication is array
provenance. In the special case where both arrays have static array
corresponding to the empty array, the data actually corresponds to the
BArrLikeE bits, so we need to check that. Update the comment to
indicate this isn't just an optimization.

Reviewed By: markw65

Differential Revision: D19374431

fbshipit-source-id: a803acde40fd8c613268de2b6d58877e3765a50a