tyvar_occurrences was not maintained properly when shortening paths
Summary: There are times when we call `add var ty` on a var which is already bound to another type, for example when we do path shortening in `Typing_inference_env.get_type`. In those cases, the tyvar_occurrences where not updated properly, as they would keep the previous variable binding. This diff fixes that.

Differential Revision: D19346792

fbshipit-source-id: 9b9c912d8d9ebf4ece5a4f581a7b15fe2fb94f72