Fix operations between static arrays and non-static specialized arrays in HHBBC
Summary:
It's an invariant in HHBBC that specialized static arrays
cannot contain non-static types (since such arrays cannot exist at
runtime). Most operations maintain this invariant for free, but it can
be broken by intersection. If you have a non-static specialized array,
and intersect it with a static (non-specialized) array, the result
will be static, but with the specialization of the non-static
array. If the specialization contained a type which isn't allowed in
static arrays, we'll fail the invariant check. Note that this can only
happen if the static array has no specialization. If it does, we'll
intersect both specializations, which will maintain the invariant
properly.

Fix this by checking for this situation (intersection of a non-static
specialized array with a static non-specialized array). Introduce a
new helper function called remove_counted() which makes a type static,
including any specialization information recursively. Use this to fix
the specialization during this kind of intersection.

A similar problem happens with the couldBe operator. Fix it in a
similar way, with a new must_be_counted() function.

Reviewed By: jano

Differential Revision: D19326208

fbshipit-source-id: 638d4147fb8c76e5c4d369e024fe910f68c1c922