Restore HHI root if deleted
Summary:
Some cleanup systems may delete files in `/tmp` if they're too old. Before IDE requests, restore it to disk if it no longer exists.

There are a few options:

* Kill the client if the HHI directory is no longer there.
* Touch the HHI directory every so often, so that it doesn't get cleaned up.
* Re-extract the HHI directory if it's no longer there.

The simplest solution is actually to re-extract the HHI directory. There's not a way for the serverless IDE to indicate that it's purposefully dying at the moment, and we would need to set up an asynchronous loop to touch the HHI directory every so often. It's also pretty robust to do so.

Reviewed By: ljw1004

Differential Revision: D19315429

fbshipit-source-id: 1e7507df4202c0ea9ed266e17971ea385b2bf76b