Use counted array types in JIT if RepoAuthType contains a counted type
Summary:
HHBBC doesn't infer counted array types (and we never exported them in
RepoAuthType anyways), but if the RepoAuthType contains a counted type
(like Object or Resource), we know the array must be counted. When
turning the RepoAuthType into a Type, check for this and generate the
correct array type.

Reviewed By: paulbiss

Differential Revision: D19397204

fbshipit-source-id: 6564bebacd74feb246bc3f7c103c1c0740732a68