Fix OOB error in simplifyLdClsMethod
Summary:
When we profile a single function target for a method call, we load the function at the current class's slot for that method using LdClsMethod, then compare it to our profiled target. We have a simplify pass that converts LdClsMethod into a constant result if the class and index are both constant. This simplify pass doesn't do a bounds check on the class's method vector length. It's a little bit hard to trigger this bug, because if the class itself is unique, we'll check the current class against the profiled class instead of loading the function. As a result, in the unit test for this bug, I defined a class ("HasFoo") with two methods (causing the OOB when we index into it at index 1) and with two subclasses (so the method will be unique but the class will not).

The difference between the "logical index" (between [0, m_funcVecLen)), the "physical index" = -(logical index + 1) and the "physical offset" = sizeof(LowFuncPtr) * (physical index) is confusing, so I changed this op from taking a physical index to taking a logical index.

The code we originally JIT for this optimization is correct, because it emits a bounds check inline, using LdFuncVecLen. However, there's no simplify pass for LdFuncVecLen, and I don't want to add one right now, because it's not current exposed through the class API. (Note that a class's "func vector" may be larger than its number of methods, because of traits, etc.)

Reviewed By: ricklavoie

Differential Revision: D19397719

fbshipit-source-id: 090fcdabbf5a8cc381c8c3323ff8f1e1ad76545a