Check that XHP children declarations do not refer to undefined types
Summary:
[xhp children](https://docs.hhvm.com/hack/XHP/extending#children) declarations are currently not typechecked. In that, you can add random stuff there and hh won't complain.

Before HackC, only XHP-specific types (such as element names, categories, `pcdata`, `any`) were permitted, not arbitrary names.

This diff will throw unbound name errors when xhp children are not declared.

This diff does not verify that children match the declaration; for example, this is an error:

```
class :foo {
  children (idonotexist);
}
```

but this is not:

```
class :foo {
  children (:bar);
}
class :bar {}
class :baz {}

$_ = <foo><baz /></foo>;
```

... despite the fact that `:baz` is not a valid child of `:foo`

This is still not completely correct. Type aliases with generics and non final classes with generics should not be allowed as the type structure is not properly encoded in the bytecode. However, due to time/utility tradeoff, punting on both of those items. We want to reevaluate the existence of the `children` feature.

Hack breaking change - we will no longer accept `PCDATA`, `PcDatA`, or cased variations of `pcDaTa` or other existing classes as valid.

Reviewed By: Wilfred

Differential Revision: D19353526

fbshipit-source-id: 6d39de731609ea6203ef06a7faa959de6c8ede1a