Make inout assignment error position independent of the origins of the types being checked
Summary:
At the end of a function body, the types of the locals that correspond to the inout parameters are fetched and checked for compatibility with their expected type. However, this uses the standard unification error and the primary position comes from the type in the subtype position. This creates situations where the primary position of an error can be in another file.

This diff changes the position to the inout parameter, though the closing brace of the function is an equally valid position based on when we do this check.

Reviewed By: periodic1236

Differential Revision: D19358356

fbshipit-source-id: 17e328f23d6e9af764bfedd9cb5c79fe42a73b54