Mark FFI functions unsafe
Summary:
I always intended for these to be marked unsafe, but wasn't sure how to forbid `unsafe` blocks within the body (`$code`) if the extern function was marked unsafe. I had tried wrapping `$code` in a closure, but the `unsafe` keyword on the outer function definition permits unsafe code in closures within the function. Turns out all I needed to do was define an inner function rather than a closure.

Here's an example to explain in more detail. If we have this macro invocation:

```
ocaml_ffi! {
    fn distance(a: (i64, i64), b: (i64, i64)) -> f64 {
        let dx = b.0 - a.0;
        let dy = b.1 - a.1;
        let sum = dx.pow(2) + dy.pow(2);
        (sum as f64).sqrt()
    }
}
```

We previously expanded it to this function definition:

```
#[no_mangle]
pub extern "C" fn distance(a: usize, b: usize) -> usize {
    ocamlrep_ocamlpool::catch_unwind(|| {
        use ocamlrep_ocamlpool::OcamlRep;
        let a = unsafe { <(i64, i64)>::from_ocaml(a).unwrap() };
        let b = unsafe { <(i64, i64)>::from_ocaml(b).unwrap() };
        let result: f64 = {
            let dx = b.0 - a.0;
            let dy = b.1 - a.1;
            let sum = dx.pow(2) + dy.pow(2);
            (sum as f64).sqrt()
        };
        unsafe { ocamlrep_ocamlpool::to_ocaml(&result) }
    })
}
```

But this is a problem--the function contains unsafe blocks, but does nothing to uphold their invariants (the arguments must be valid OCaml values, and no other threads may interact with the OCaml runtime while converting the return value). It is easy to cause a segfault using only safe Rust code:

```
// These arguments are supposed to be OCaml tuples. If we pass nullptr instead,
// we'll attempt to dereference null and segfault.
distance(0, 0);
```

Instead, we should mark the entire function unsafe (forbidding its use in safe Rust), and only invoke it from OCaml. The invariants we must satisfy for those unsafe blocks should then be upheld--a well-formed OCaml program will only pass valid OCaml values, and won't attempt to interact with the runtime from multiple threads.

However, if *all* we do is mark the entire function unsafe, we have another problem--unsafe code is permitted in the entire body of the `distance` function! If we change it to use unsafe code:

```
ocaml_ffi! {
    fn distance(a: (i64, i64), b: (i64, i64)) -> f64 {
        let dx = b.0 - a.0;
        let dy = b.1 - a.1;
        let sum = dx.pow(2) + dy.pow(2);
        // unsafe, and wrong--but no unsafe block!
        std::mem::transmute::<i64, f64>(sum).sqrt()
    }
}
```

Then our naive macro might expand it to

```
#[no_mangle]
pub unsafe extern "C" fn distance(a: usize, b: usize) -> usize {
    ocamlrep_ocamlpool::catch_unwind(|| {
        use ocamlrep_ocamlpool::OcamlRep;
        let a = <(i64, i64)>::from_ocaml(a).unwrap();
        let b = <(i64, i64)>::from_ocaml(b).unwrap();
        let result: f64 = {
            let dx = b.0 - a.0;
            let dy = b.1 - a.1;
            let sum = dx.pow(2) + dy.pow(2);
            // legal, due to the containing unsafe fn!
            std::mem::transmute::<i64, f64>(sum).sqrt()
        };
        ocamlrep_ocamlpool::to_ocaml(&result)
    })
}
```

I don't want this--if we use unsafe functions in the body, we should be required to call it out with an unsafe block.

This diff solves this problem by expanding the original macro invocation to this instead:

```
#[no_mangle]
pub unsafe extern "C" fn distance(a: usize, b: usize) -> usize {
    ocamlrep_ocamlpool::catch_unwind(|| {
        fn inner(a: (i64, i64), b: (i64, i64)) -> f64 {
            let dx = b.0 - a.0;
            let dy = b.1 - a.1;
            let sum = dx.pow(2) + dy.pow(2);
            (sum as f64).sqrt()
        }
        use ocamlrep_ocamlpool::OcamlRep;
        let a = <(i64, i64)>::from_ocaml(a).unwrap();
        let b = <(i64, i64)>::from_ocaml(b).unwrap();
        let result = inner(a, b);
        ocamlrep_ocamlpool::to_ocaml(&result)
    })
}
```

Because the `extern fn` is now marked `unsafe`, we cannot cause a segfault using safe code as before. Because the body is now in an inner function which is **not** marked `unsafe`, we still forbid unsafe code within it. This also allows us to add `return` statements within that body (previously, they'd have caused an early-return, skipping the `to_ocaml` call).

Reviewed By: Wilfred

Differential Revision: D19356222

fbshipit-source-id: 320b0bef4dd796e30befc4129a86fac770b2de19