Move env setting from iter_with to iter
Summary:
Examining our usages of tast_check and the tast_visitor, I believe that we should be more disciplined about the way that we write tast checks.

Rather than set tast_env details in the `iter_with` class, I'd rather set the environmental items in the `iter` class - this has the effect of correctly having all classes inheriting from `iter` to also get the change. We need to duplicate this for the `iter_with_state`. This is also more consistent with the current pattern of env setting.

Reviewed By: losvald

Differential Revision: D19044954

fbshipit-source-id: e4aaa6dfcf92ea30a9f820556fdee210bfa0cc43