Attached debuggers cause segfaults with timeout and user timeout combined
Summary:
Returning `TimedOutFlag` surprise flag means a user timeout callback should happen.
We have an if/elseif/else that makes sure that never happens unless we want the callback to happen.
Except when the debugger is on, right now if the debugger is on we skip that if altogether and accidentally return a signal that a callback should happen.

Differential Revision: D19419859

fbshipit-source-id: f5edd58dd5f3fe5da74421c833bfa9c1350938f3