Use class NASTs instead of class decls to generate class definitions
Summary:
At the moment, to generate a class definition we look up the class decl in the decl heap by the class name and generate a definition off of that.  We are forced to do some extra computations to reconstruct the list of direct ancestors because the decl data structure contains only information about *all* ancestors.  Furthermore, because the ancestors are stored in an `SMap`, we lose the order in which they are specified in the `extends` and `implements` clauses in the source code.  Unfortunately, this order is important (at least, with the current implementation of typing) and affects type checking.

For example:
```
interface N {
  abstract const type T as num;
}

interface I {
  abstract const type T as int;
}

interface NI extends N, I {}
interface IN extends I, N {}
```
Here `IN` type checks but `NI` produces a type error
```
Class NI does not correctly implement all required members  (Typing[4340])
```
We have way too many examples of this patter to ignore it.

Here I'm proposing to generate class definitions off of their NAST nodes.  In this diff I use this approach to generate the ancestors but I envision doing pretty much everything off of NASTs rather than decls.

Reviewed By: andrewjkennedy

Differential Revision: D19392901

fbshipit-source-id: bf3a34c583f28d59d4cc1df807cbcc41bf3cac79