Improve ArrLikeMap in HHBBC
Summary:
ArrLikeMap in HHBBC represents arrays which have a complete known set
of keys (in that precise order). If at any point, we no longer know
the complete set of keys, we fall back to the ArrLikeMapN
representation (which is just a key/value type tuple). However, this
can discard a considerable amount of information.

Improve the representation by adding two new type fields to
ArrLikeMap. These types represent an upper bound on the keys or values
of the elements in the array that do not have known static keys. The
keys section of ArrLikeMap continue to represent the known static keys
of the array (starting from the beginning of the array). If the new
type fields are Bottom, these represent the complete set of keys in
the array (as before). If they are not, however, it means that after
the set of keys, there's some possible set of array entries with those
key/value types. Since these types are upper-bounds, it does not imply
there's always extra array entries.

This is advantageous for situations where you union together two
ArrLikeMaps that share a common prefix of keys. Before we'd have to
drop to ArrLikeMapN, but now we can preserve the common prefix.

Add the types to HHBBC and update all the type operations to
update/use it appropriately.

I originally tried an even more precise analysis (preserving specific
optional keys), but this lead to excessive HHBBC runtimes due to
pathological cases causing an exponential blowup in representation
size. Keeping a single key/value type pair avoids this blowup.

Reviewed By: jano

Differential Revision: D19151479

fbshipit-source-id: 357c3dabd3dbdc037137972f62838d37719f2f7e