Use function NASTs to generate function and method declarations
Summary: Extract the information required to generate function and method declarations from the corresponding NAST nodes rather than from decls.  In particular, this allows us to more accurately generate constructors and `__Override` attributes.  (The latter actually forces us to track more dependencies: when a method is marked as `__Override`, it must be an override of something, so we need to add a dependency on the parent method.  I don't know of a good way to do that, so I'm traversing the list of all ancestors and add dependencies on those that have methods with the given name.)

Reviewed By: Wilfred

Differential Revision: D19410436

fbshipit-source-id: bcdea965b63c9957310d5fdf8c55bde50e622710