Ensure that the has_xhp_keyword field is set correctly (#8644)
Summary:
This is a bugfix on top of https://github.com/facebook/hhvm/issues/8638. In that PR we changed the XHP keyword from a class `kinds` to be a separate optional keyword before `class`. However the lowerer.rs line below remained the same, fixing to read the xhp keyword instead of attempting to find it in kinds.
Pull Request resolved: https://github.com/facebook/hhvm/pull/8644

Reviewed By: Wilfred

Differential Revision: D19450589

Pulled By: jjergus

fbshipit-source-id: b6b10e6091e2f4c8df2fc569c81f86267dddf152