Enabling single request strobelight profiling
Summary: This adds a cleaner path for strobelight profiling that sets only a single surprise flag on a specific thread. bpf_send_signal_thread is a new (pending) kernel feature that will use this new signal. The original signal and signal handler remain in place because they are useful for dumping all current request stacks, which is something we are also interested in.

Reviewed By: binliu19

Differential Revision: D19395215

fbshipit-source-id: 14cdc555fe07fa359c987f9af428a4c47765f51f