still add type var to inference env even when outside of a tyvar scope
Summary:
It might be that we create type variables outside of a tyvar scope. This is usually a programming error and we should probably throw an exception, but we can at least still add the type variable in the environment.

Context: with global inference, we've seen that there seems to be calls to functions like add_tyvar_lower_bound to either non-existent variables or variables already bound. Ultimately we want to raise an exception in those cases.
This diff avoids one case where we forget to create a type variable.

Differential Revision: D19448447

fbshipit-source-id: 5baa32c47183277d33d857df2bd3e7bad3eaec1f