Move parser_env.rs from parser_rust to parser_core_types
Summary:
In D19440902, I limit the visibility of main parser crate to prevent people from shooting themselves in the foot. This requires moving parts that are part of public API away from the soon to be private TARGET.

I perpetuate here (and in subsequent diffs) the re-exporting of crates where they previously used to be to limit amount of renaming necessary for now, and delay it until it's needed for dependency management reasons (like in D19443437).

(Note: this ignores all push blocking failures!)

Reviewed By: shiqicao

Differential Revision: D19439811

fbshipit-source-id: bb616c836d800d213da04d11509b2fa73dd6cfcb