Move parser:parser_rust to parser/core:parser and limit it's TARGET visibility
Summary:
The `parser` data structure is dangerous - every crate that instantiates it will take a looong time to compile. It's also not user friendly - constructing new parser requires defining a very complicated type involving multiple generics and trait bounds.

To solve both of those problems I limit the visibility of this crate to a whitelist (see `visibility = ...` in `parser/core/TARGETS`), and all the subsequent diffs up to D19443963 are about trimming down this whitelist to a minimal set where each parser used anywhere in codebase is instantiated only once, in a crate with minimal dependencies, and which exposes a non-generic API preventing further spill of generic cost.

Because in buck sibling TARGETS are always visible to each other, I also need to move the core parser data to a subdir with it's own TARGET.

(Note: this ignores all push blocking failures!)

Reviewed By: shiqicao

Differential Revision: D19440902

fbshipit-source-id: 72fe6884339df8b05a3ebbf219016003bdcad8f5