Remove rust_parser_ffi dependency on core parser
Summary:
I want `parse` function to not instantiate new parser, but re-use one of the APIs defined in this diff stack. This requires adding one more generic to it defining the function which should be used (it's pretty gross).

This concludes cleaning of the core parser whitelist (only things in /api/ depend on it now), so I also add a comment warning against adding more things there.

(Note: this ignores all push blocking failures!)

Reviewed By: shiqicao

Differential Revision: D19443963

fbshipit-source-id: 76ca354eaa97cd76f4fadd12b3579c55d529e5c3