Split the slow/rx/body tests.
Summary:
Split the slow/rx/body tests to have one kind of syntax per test, to make it easier to add parser errors for them (we'll be able to add errors one at a time if needed, whereas with the way the tests are structured now it would require converting a whole category of these at a time).

In order to facilitate this, make the config for the slow/rx/body directory turn the body verification errors up to fatal and add .verify files for all the tests; this way the tests don't run after failing verification. Since it is now the default, instead of having one test to test fatal mode (fail.php) have one test that tests the non-fatal mode (warn.php).

Reviewed By: DavidSnider

Differential Revision: D19454208

fbshipit-source-id: d53763c9b9c0d513046e08b2deb94b89851dc14f