log semantic version
Summary:
Currently, every scuba sample has a "Version" field which is the build hash.

When reading logs, I really just want to see what the semantic version is.

The problem is that "semantic version" doesn't exist intrinsically to any of our binaries. The only things that exist are
1. the "version=" line in .hhconfig
  * hh_server reads that version= line and uses it to enable or disable features in hh.conf
  * clientlsp reads that version= line, doesn't interpret it, but notices if ever its contents change
2. the directory path in which hack is launched
  * find_hh.sh reads the version= line and uses that to pick which directory path to launch hack from

So what would it ever mean to log semantic version information to our telemetry?

Here, I picked the interpretation that we will log the version which we picked up from hhconfig. That after all is the same interpretation used within hh_server for interpreting hh.conf, so is a good one to standardize upon.

There are many tools which don't even bother reading the version= line from hhconfig, or which initialize telemetry-logging prior to having read the field. For all of them I've just left telemetry about hhconfig_version to be empty.

But for every tool which does read and use the version= line, this diff makes it now write that hhconfig_version to telemetry.

Reviewed By: 2BitSalute

Differential Revision: D19459943

fbshipit-source-id: d4f4ee53024a1133170b2ac123195bcb6c135774