Stricter type checks for pointer types
Summary:
Some IR ops, like ElemArrayD and friends, take "LvalToCell" state that the lval must point to an array, and may even take the target type as a separate type parameter. Our type system is capable of expressing constraints on pointer target types, so we can eliminate the text documentation and the additional type parameter and just rely on the existing checking.

Making this change requires some improvements to our type-tracking for the member base register in FrameState and the addition of type parameters to LdLocAddr / LdStkAddr.

Finally, we can also tighten up a few ops like "InitPackedLayoutArray" that take S(AK(Packed),Vec) as a source. Because of representation issues, the union type AK(Packed)|Vec is equivalent to Arr|Vec, losing type information. Instead, we distinguish between constraining an IR op's source to be *a subtype of any of a list of types* or *a subtype of a union type*, where the former is strictly stronger. This distinction gives us a nice way to represent the true type constraint of ops like ArrayIdx, which can take an int or a string, but not a union of the two.

Reviewed By: jjgriego

Differential Revision: D19403780

fbshipit-source-id: c0254ca42539edd3d134b17bf59be646eb62fc69