Devirtualize CallSpec::array when possible
Summary:
When we know the kind of an array, we should emit a direct call to one if its array funcs rather than a virtual call that loads its kind.

We modify CallSpec::array to do this devirtualization for us. It now takes a jit::Type for the array so that callers can't miss this optimization, as well as a virtual method to fall back on if we don't know the kind. CallSpec::array is template in such a way that the direct calls and the method call have to have parallel argument types, albeit with the direct call taking the ArrayData* itself as an additional first parameter.

Reviewed By: ricklavoie, mofarrell

Differential Revision: D18778554

fbshipit-source-id: a746ea13ea232386d001fb9c2751ea5d520dda08