Eliminate (Dict)?AddElem(Int|Str)Key
Summary:
Putting this change up for review again, stacked on top of the array devirtualization changes.

We only use AddElem right after doing a NewDArray, and NewDArray returns a known MixedArray. Now that we're devirtualizing ArraySet, it looks like a branch miss win to use it with a known array kind. I'll confirm that it's a win in a load test before landing the change this time.

Reviewed By: mofarrell

Differential Revision: D19466139

fbshipit-source-id: 1e2f0abc2c4802c01acb85a47e2320df597c5b1d