Extract a function for retrieving method NASTs and memoize it
Summary: Previously, finding a method by its name and the class name required traversing the list of methods, so that getting the methods of a given class was linear time.  Here I extract this into a helper function and memoize it (on the first call the list of methods of the class with the supplied name is converted into a map and is cached).

Reviewed By: andrewjkennedy

Differential Revision: D19427991

fbshipit-source-id: 7d5d2a7cde6c23b77fd9e075827b2f8ae1b23266