Memoize functions for retrieving function and class NAST nodes by name
Summary: What it says on the tin.  I'm not sure if `Ast_provider` already caches NAST nodes or does reparsing of the source on each call, and I couldn't be bothered to check.  I'd like to be able to call `get_fun_nast` and `get_class_nast` from different functions and expect them to be reasonably cheap.

Reviewed By: andrewjkennedy

Differential Revision: D19428916

fbshipit-source-id: b12b749788580906bdfedf233e316285fa691971