Treat constructors as regular methods
Summary: I don't think we need to treat constructors specially anymore.  We generate a constructor for a class if that class has an explicit constructor defined in the source code.  Otherwise the constructor is either default (and doesn't need to be generated) or is inherited from an ancestor, in which it is defined explicitly and that constructor will be generated.

Reviewed By: andrewjkennedy

Differential Revision: D19449089

fbshipit-source-id: c179050ea1e183e58d4ef00aa5d6247786cbb94b