Ignore non-built-in user attributes with parameters
Summary:
It turns out user attributes with parameters are pretty common.  To be able to print them, we need an AST printer.  For now, let us drop user attributes with parameters.

Furthermore, because we take the list of user attributes from a NAST node, for example, when we generate a stub for one of the dependency functions / methods, we don't re-type check that function / method (we just look at its type signature), and so its non-built-in user attributes are not added as dependencies.  If we print all user attributes (without parameter), including the non-built-in ones, we may end up with code that references non-existing classes.  We don't want that, so let us further restrict the set of printable user attributes to built-in ones only (still without parameters).

Reviewed By: andrewjkennedy

Differential Revision: D19470654

fbshipit-source-id: dbce34ba553e17a714eecec7a4ba05d5a3d63c94