Distinguish equality vs relational comparison notices for v/darray
Summary:
This diff changes the error string for notices that are raised for expressions like `varray[] < darray[]` and `varray[] === darray[]` so that the errors strings are different.

Being able to distinguish these two cases by the error string will allow us to treat them differently in the error elevator.

When enforcing the specialization of varray and darray, we will want to have relational comparisons result in an exception while equality comparisons should change behavior (always return false for varray vs darray). We may choose to use the error elevator to raise exceptions for just the former and subsequently use a config change to change the behavior in the runtime.

Reviewed By: jthemphill

Differential Revision: D19507325

fbshipit-source-id: 999430cca3c95591249db427c9912aeca30aef94