Implement Display for Pos
Summary:
Add an impl of Display for Pos. It displays the filename and the range using line/column numbers.

I'm not sure if we want this to be the default implementation of Display for Pos or not. There are several ways we display positions (consider, for example, [the way they are displayed for typecheck test output](https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/test/typecheck/do2.php.exp?commit=e48c48295afbbdb794eb5efc5d3df89a4d9535ba&lines=1)). We could always move this implementation to a wrapper struct for this specific display format if we decide against keeping this as the default (as the next diff in the stack does for `Error_`).

Reviewed By: losvald

Differential Revision: D19458425

fbshipit-source-id: 7d76766f42b44a5246a8bed400970a23a1752612