Support functions
Summary:
We currently only support reading the fileinfo table and consts table. This diff adds support for reading the funs table.

I'd like to look into reducing the redundancy between the different symbol kinds in a near-future diff, but for now, the funs module is a copy-paste of the consts module.

Reviewed By: 2BitSalute

Differential Revision: D19522847

fbshipit-source-id: 2faf2068cf86b530153df93d3ff38c68cd75530b