don't error on registerCapability
Summary:
We send a client/registerCapability request, to ask for file change notifications.

The client sends back a response.

But we never got around to coding anything to parse that response. Consequently, we report an error that looks scary in the logs:
```
[Trace - 9:38:41 PM] Received request 'telemetry/event - (10000000000004)'
Result: {
  "jsonrpc": "2.0",
  "method": "telemetry/event",
  "params": {
    "type": 1,
    "message": "Lsp.Error.Parse(\"Don't know how to parse LSP response client/registerCapability\"), from_client\nRaised at file \"/data/users/ljw/fbsource/fbcode/hphp/hack/src/utils/lsp/lsp_fmt.ml\", line 1358..."
  }
}
```

This diff adds ability to parse the response (which has empty content) so as to avoid the scary log messages.

Reviewed By: arxanas

Differential Revision: D19531019

fbshipit-source-id: a81da97f0f497806903a9e1a1520b03059ffa32a